+++ ruma-identifiers-validation/src/device_key_id.rs
@@ -3,7 +3,7 @@ use std::{num::NonZeroU8, str::FromStr};
 use crate::{crypto_algorithms::DeviceKeyAlgorithm, Error};
 
 pub fn validate(s: &str) -> Result<NonZeroU8, Error> {
-    let colon_idx = NonZeroU8::new(s.find(':').ok_or(Error::MissingDeviceKeyDelimiter)? as u8)
+    let colon_idx = NonZeroU8::new(s.find(':').ok_or(Error::MissingDelimiter)? as u8)
         .ok_or(Error::UnknownKeyAlgorithm)?;
 
     DeviceKeyAlgorithm::from_str(&s[0..colon_idx.get() as usize])