+++ ruma-identifiers-validation/src/error.rs
@@ -22,17 +22,12 @@ pub enum Error {
     /// The ID exceeds 255 bytes (or 32 codepoints for a room version ID).
     MaximumLengthExceeded,
 
-    /// The ID is missing the colon delimiter between localpart and server name.
+    /// The ID is missing the colon delimiter between localpart and server name, or between key
+    /// algorithm and key name / version.
     MissingDelimiter,
 
-    /// The ID is missing the colon delimiter between key algorithm and device ID.
-    MissingDeviceKeyDelimiter,
-
-    /// The ID is missing the colon delimiter between key algorithm and version.
-    MissingSigningKeyDelimiter,
-
     /// The ID is missing the correct leading sigil.
-    MissingSigil,
+    MissingLeadingSigil,
 
     /// The key algorithm is not recognized.
     UnknownKeyAlgorithm,
@@ -46,10 +41,8 @@ impl Display for Error {
             Error::InvalidKeyVersion => "key ID version contains invalid characters",
             Error::InvalidServerName => "server name is not a valid IP address or domain name",
             Error::MaximumLengthExceeded => "ID exceeds 255 bytes",
-            Error::MissingDelimiter => "colon is required between localpart and server name",
-            Error::MissingDeviceKeyDelimiter => "colon is required between algorithm and device ID",
-            Error::MissingSigningKeyDelimiter => "colon is required between algorithm and version",
-            Error::MissingSigil => "leading sigil is incorrect or missing",
+            Error::MissingDelimiter => "required colon is missing",
+            Error::MissingLeadingSigil => "leading sigil is incorrect or missing",
             Error::UnknownKeyAlgorithm => "unknown key algorithm specified",
         };