+++ ruma-identifiers/src/room_id_or_room_alias_id.rs
@@ -166,7 +166,10 @@ mod tests {
 
     #[test]
     fn missing_sigil_for_room_id_or_alias_id() {
-        assert_eq!(RoomIdOrAliasId::try_from("ruma:example.com").unwrap_err(), Error::MissingSigil);
+        assert_eq!(
+            RoomIdOrAliasId::try_from("ruma:example.com").unwrap_err(),
+            Error::MissingLeadingSigil
+        );
     }
 
     #[cfg(feature = "serde")]